Refactor folders 12-14: restructure PBD cloth, mesh, and PBF simulators
