+++ 12_pbd_cloth/constraints/__init__.py
@@ -1,0 +1,7 @@
+from .stretching import solve_stretching_constraints
+from .bending import solve_bending_constraints
+
+__all__ = [
+    'solve_stretching_constraints',
+    'solve_bending_constraints',
+]
