+++ 14_pbf/constraints/__init__.py
@@ -1,0 +1,10 @@
+from .density import solve_density_constraints
+from .vorticity import calculate_vorticity, apply_vorticity_confinement
+from .viscosity import apply_xsph_viscosity
+
+__all__ = [
+    'solve_density_constraints',
+    'calculate_vorticity',
+    'apply_vorticity_confinement',
+    'apply_xsph_viscosity',
+]
