+++ pyproject.toml
@@ -1,0 +1,11 @@
+[project]
+name = "solid-sim-tutorial"
+version = "0.1.0"
+description = "Add your description here"
+readme = "README.md"
+requires-python = ">=3.13"
+dependencies = [
+    "matplotlib>=3.10.3",
+    "numpy>=2.3.1",
+    "scikit-learn>=1.7.0",
+]
